requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-pptx==0.6.21
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psutil==5.9.5
Pillow==10.0.1
XlsxWriter==3.1.9
gunicorn==21.2.0
matplotlib==3.7.2
pandas==2.0.3
numpy==1.24.3
seaborn==0.12.2